% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_links.R
\name{append_links}
\alias{append_links}
\title{Append links to cv printer attributes}
\usage{
append_links(cv, new_links)
}
\arguments{
\item{cv}{object of class `cv_printer` created by \code{\link{new_cv_printer}}}

\item{new_links}{Character array of new links to append to links attribute.}
}
\value{

}
\description{
Helper function to append links to a `cv_printer` object's `links` attribute.
}
