% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_cv_printer.R
\name{new_cv_printer}
\alias{new_cv_printer}
\title{Setup CV printer object}
\usage{
new_cv_printer(cv_data, pdf_mode = TRUE)
}
\arguments{
\item{cv_data}{List of dataframes containging info for CV}

\item{pdf_mode}{Boolean of if document is being exported as pdf and thus
links should be stripped for placement at end.}
}
\value{

}
\description{
Setup CV printer object
}
