% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_contact_info.R
\name{print_contact_info}
\alias{print_contact_info}
\title{Print contact info}
\usage{
print_contact_info(cv)
}
\arguments{
\item{cv}{object of class `cv_printer` created by \code{\link{new_cv_printer}}}
}
\value{

}
\description{
Prints out an icon-bulleted list of contact info as supplied in `contact_info` slot of cv data.
}
