% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_links.R
\name{print_links}
\alias{print_links}
\title{Print stripped links to section}
\usage{
print_links(cv)
}
\arguments{
\item{cv}{object of class `cv_printer` created by \code{\link{new_cv_printer}}}
}
\value{

}
\description{
Prints all the links that were stripped from various text as cv print object
was used to a single section with the link footnote identifier next to each
link.
}
\examples{

# Function to get sample csvs from package data store.
sample_data <- function(name) system.file(paste0('sample_csvs/', name), package = "cvdown")

# Run function on these sample datasets
data <- gather_data_from_csvs(positions_loc = sample_data('positions.csv'),
                              skills_loc = sample_data('language_skills.csv'),
                              text_blocks_loc = sample_data('text_blocks.csv'),
                              contact_info_loc = sample_data('contact_info.csv'))

# Setup CV printer
printer <- new_cv_printer(data, pdf_mode = TRUE)

# Print intro text block with links in it
printer <- print_text_block(printer, 'intro')

# Print out links that were extracted
print_links(printer)

}
