% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_section.R
\name{print_section}
\alias{print_section}
\title{Print position section}
\usage{
print_section(cv, section_id)
}
\arguments{
\item{cv}{object of class `cv_printer` created by \code{\link{new_cv_printer}}}

\item{section_id}{Name of section to be printed as encoded by the column `section` in positions dataframe.}
}
\value{

}
\description{
Print position section
}
