% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_skill_bars.R
\name{print_skill_bars}
\alias{print_skill_bars}
\title{Print skill bars}
\usage{
print_skill_bars(
  cv,
  out_of = 5,
  bar_color = "#969696",
  bar_background = "#d9d9d9"
)
}
\arguments{
\item{cv}{object of class `cv_printer` created by \code{\link{new_cv_printer}}}

\item{out_of}{What is the skill value out of?}

\item{bar_color}{Color of filled in skill bar}

\item{bar_background}{Color of background (unfilled in part) of skill bar.}
}
\value{

}
\description{
Writes out the skill bars as divs with varying widths based on skill level
}
\examples{

# Function to get sample csvs from package data store.
sample_data <- function(name) system.file(paste0('sample_csvs/', name), package = "cvdown")

# Run function on these sample datasets
data <- gather_data_from_csvs(positions_loc = sample_data('positions.csv'),
                              skills_loc = sample_data('language_skills.csv'),
                              text_blocks_loc = sample_data('text_blocks.csv'),
                              contact_info_loc = sample_data('contact_info.csv'))

# Setup CV printer
printer <- new_cv_printer(data, pdf_mode = TRUE)

# Print skill bars
printer <- print_skill_bars(printer)


}
