% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_text_block.R
\name{print_text_block}
\alias{print_text_block}
\title{Print text block}
\usage{
print_text_block(cv, label)
}
\arguments{
\item{cv}{object of class `cv_printer` created by \code{\link{new_cv_printer}}}

\item{label}{value in the `label` column of `text_blocks` dataframe for row that holds text desired.}
}
\value{

}
\description{
Prints arbitrary block of text from the `text_blocks` section of CV data
}
