% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_links.R
\name{extract_links}
\alias{extract_links}
\title{Extract links from text}
\usage{
extract_links(text, cv)
}
\arguments{
\item{text}{Text containing markdown formatted links.}

\item{cv}{Object of class `cv_printer` that has `links` attribute containing
character vector of links. Used to determin what to start footnote
reference number at.}
}
\value{

}
\description{
Helper function that takes a block of text, extracts all markdown formatted
links and returns the text with all links stripped out and replaced by
footnote references and an array of the new links to add to `cv_printer`
attributes.
}
