% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_data_from_googlesheets.R
\name{gather_data_from_googlesheets}
\alias{gather_data_from_googlesheets}
\title{Gather CV data from a google sheet}
\usage{
gather_data_from_googlesheets(positions_sheet_loc = NULL, public_sheet = FALSE)
}
\arguments{
\item{public_sheet}{Info on if the sheet is public or not. Usually CV's are
public so your sheet should be to. Make a sheet public by going to "share"
and saying "anyone with link can view".?}
}
\value{

}
\description{
Loads all cv data into a single list that is taken by the new_cv_printed
function and used to generate your data-driven CV!
}
