% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_cv_printer.R
\name{new_cv_printer}
\alias{new_cv_printer}
\title{Setup CV printer object}
\usage{
new_cv_printer(
  cv_data,
  pdf_mode = TRUE,
 
    position_template = "\\n                            ### {title}\\n                            \\n\\n\\n                            {loc}\\n                            \\n\\n\\n                            {institution}\\n                            \\n\\n\\n                            {timeline}\\n                            \\n\\n\\n                            \\n\\n\\n                            {description_bullets}\\n                            \\n\\n "
)
}
\arguments{
\item{cv_data}{List of dataframes containging info for CV}

\item{pdf_mode}{Boolean of if document is being exported as pdf and thus
links should be stripped for placement at end.}

\item{position_template}{Glue tempate string that is used to build each position entry.}
}
\value{

}
\description{
Setup CV printer object
}
\examples{

# Function to get sample csvs from package data store.
sample_data <- function(name) system.file(paste0('sample_csvs/', name), package = "cvdown")

# Run function on these sample datasets
data <- gather_data_from_csvs(positions_loc = sample_data('positions.csv'),
                              skills_loc = sample_data('language_skills.csv'),
                              text_blocks_loc = sample_data('text_blocks.csv'),
                              contact_info_loc = sample_data('contact_info.csv'))

# Setup CV printer
printer <- new_cv_printer(data, pdf_mode = TRUE)

# Print contact info
printer <- printer \%>\% print_contact_info()

}
