% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_contact_info.R
\name{print_contact_info}
\alias{print_contact_info}
\title{Print contact info}
\usage{
print_contact_info(cv)
}
\arguments{
\item{cv}{object of class `cv_printer` created by \code{\link{new_cv_printer}}}
}
\value{

}
\description{
Prints out an icon-bulleted list of contact info as supplied in `contact_info` slot of cv data.
}
\examples{

# Function to get sample csvs from package data store.
sample_data <- function(name) system.file(paste0('sample_csvs/', name), package = "cvdown")

# Run function on these sample datasets
data <- gather_data_from_csvs(positions_loc = sample_data('positions.csv'),
                              skills_loc = sample_data('language_skills.csv'),
                              text_blocks_loc = sample_data('text_blocks.csv'),
                              contact_info_loc = sample_data('contact_info.csv'))

# Setup CV printer
printer <- new_cv_printer(data, pdf_mode = TRUE)

# Print contact info
printer <- print_contact_info(printer)


}
