% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_links.R
\name{print_links}
\alias{print_links}
\title{Print stripped links to section}
\usage{
print_links(cv)
}
\arguments{
\item{cv}{object of class `cv_printer` created by \code{\link{new_cv_printer}}}
}
\value{

}
\description{
Prints all the links that were stripped from various text as cv print object
was used to a single section with the link footnote identifier next to each
link.
}
