% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_section.R
\name{print_section}
\alias{print_section}
\title{Print position section}
\usage{
print_section(cv, section_id)
}
\arguments{
\item{cv}{object of class `cv_printer` created by \code{\link{new_cv_printer}}}

\item{section_id}{Name of section to be printed as encoded by the column `section` in positions dataframe.}
}
\value{

}
\description{
Print position section
}
\examples{

# Function to get sample csvs from package data store.
sample_data <- function(name) system.file(paste0('sample_csvs/', name), package = "cvdown")

# Run function on these sample datasets
data <- gather_data_from_csvs(positions_loc = sample_data('positions.csv'),
                              skills_loc = sample_data('language_skills.csv'),
                              text_blocks_loc = sample_data('text_blocks.csv'),
                              contact_info_loc = sample_data('contact_info.csv'))

# Setup CV printer
printer <- new_cv_printer(data, pdf_mode = TRUE)

# Print education section
printer <- print_section(printer, 'education')

}
