% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_text_block.R
\name{print_text_block}
\alias{print_text_block}
\title{Print text block}
\usage{
print_text_block(cv, label)
}
\arguments{
\item{cv}{object of class `cv_printer` created by \code{\link{new_cv_printer}}}

\item{label}{value in the `label` column of `text_blocks` dataframe for row that holds text desired.}
}
\value{

}
\description{
Prints arbitrary block of text from the `text_blocks` section of CV data
}
\examples{
# Function to get sample csvs from package data store.
sample_data <- function(name) system.file(paste0('sample_csvs/', name), package = "cvdown")

# Run function on these sample datasets
data <- gather_data_from_csvs(positions_loc = sample_data('positions.csv'),
                              skills_loc = sample_data('language_skills.csv'),
                              text_blocks_loc = sample_data('text_blocks.csv'),
                              contact_info_loc = sample_data('contact_info.csv'))

# Setup CV printer
printer <- new_cv_printer(data, pdf_mode = TRUE)

# Print a main text block
printer <- print_text_block(printer, 'intro')


}
